move DumpSlides to integration_util_test.go
